refactor + seiichi
Seiichi Hiiragi is a bullied student who gets transported to another world by a god who wants him to be a hero. He arrives in a forest, far from his classmates and civilization. He faces many dangers and hardships, but he also meets new friends and allies. He discovers a gun that can evolve him and give him amazing abilities. With this gun, he hopes to change his life and become a hero in the new world. But he also learns that the world is not as simple as he thought, and that there are secrets and mysteries waiting to be uncovered.